Use float in benchmarks

+++ benchmarks/infrastructure/config_interface.cpp
@@ -5,11 +5,11 @@
 
 class Config {
 public:
-    using DataType = std::tuple<Tensor<double>, Tensor<double>, Tensor<double>, Tensor<double>>;
+    using DataType = std::tuple<Tensor<float>, Tensor<float>, Tensor<float>, Tensor<float>>;
 
-    virtual std::tuple<Tensor<double>, Tensor<double>, Tensor<double>, Tensor<double>> LoadData(const std::string &data_path) = 0;
-    virtual NeuralNetwork<double> BuildModel() = 0;
-    virtual double GetScore(const Tensor<double> &y_true, const Tensor<double> &y_pred) = 0;
+    virtual std::tuple<Tensor<float>, Tensor<float>, Tensor<float>, Tensor<float>> LoadData(const std::string &data_path) = 0;
+    virtual NeuralNetwork<float> BuildModel() = 0;
+    virtual double GetScore(const Tensor<float> &y_true, const Tensor<float> &y_pred) = 0;
     virtual std::string GetScoreName() = 0;
 
     virtual ~Config() = default;
